refactor: Rename business directory.
