+++ src/main/java/br/com/SmallManager/infra/security/TokenService.java
@@ -1,6 +1,6 @@
 package br.com.SmallManager.infra.security;
 
-import br.com.SmallManager.domain.SystemUser;
+import br.com.SmallManager.model.SystemUser;
 import br.com.SmallManager.infra.ParameterProperties;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
