+++ src/main/java/br/com/SmallManager/repository/PeopleRepository.java
@@ -1,6 +1,6 @@
 package br.com.SmallManager.repository;
 
-import br.com.SmallManager.domain.People;
+import br.com.SmallManager.model.People;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
