+++ src/main/java/br/com/SmallManager/repository/SystemUserRepository.java
@@ -1,6 +1,6 @@
 package br.com.SmallManager.repository;
 
-import br.com.SmallManager.domain.SystemUser;
+import br.com.SmallManager.model.SystemUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
