+++ src/main/java/br/com/SmallManager/service/PeopleService.java
@@ -1,14 +1,11 @@
 package br.com.SmallManager.service;
 
-import br.com.SmallManager.domain.People;
+import br.com.SmallManager.model.People;
 import br.com.SmallManager.repository.PeopleRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 public class PeopleService {
